# docs/api/openapi.yml
{
  "openapi": "3.1.0",
  "info": {
    "title": "Whatboo API",
    "version": "2026.06.30",
    "description": "Contrato operativo para la API multi-tenant de Whatboo. Todas las rutas privadas usan el prefijo global /api/v1, JWT bearer y cabecera tenant salvo que se indique lo contrario."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API base"
    }
  ],
...
  "security": [
    {
      "bearerAuth": [],
      "tenantSlug": []
    }
  ],
  "paths": {
...
    "/customers": {
      "get": {
        "tags": ["Customers"],
        "summary": "Listar y buscar clientes del tenant actual",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/SearchQuery" },
          { "$ref": "#/components/parameters/TakeQuery" },
          { "$ref": "#/components/parameters/SkipQuery" }
        ],
        "responses": {
          "200": {
            "description": "Clientes paginados",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PaginatedCustomers" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/customers/{id}": {
      "get": {
        "tags": ["Customers"],
        "summary": "Perfil de cliente con historial de reservas",
        "parameters": [
          { "$ref": "#/components/parameters/TenantSlugHeader" },
          { "$ref": "#/components/parameters/IdPath" },
          { "$ref": "#/components/parameters/TakeQuery" },
          { "$ref": "#/components/parameters/SkipQuery" }
        ],
        "responses": {
          "200": {
            "description": "Perfil de cliente",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CustomerProfile" }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" },
      "tenantSlug": { "type": "apiKey", "in": "header", "name": "X-Tenant-Slug" },
      "onboardingToken": { "type": "apiKey", "in": "header", "name": "X-Onboarding-Token" }
    },
    "parameters": {
      "TenantSlugHeader": {
        "name": "X-Tenant-Slug",
        "in": "header",
        "required": false,
        "schema": { "type": "string" }
      },
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "string", "format": "uuid" }
      },
      "SearchQuery": {
        "name": "q",
        "in": "query",
        "required": false,
        "schema": { "type": "string" }
      },
      "TakeQuery": {
        "name": "take",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 1, "maximum": 100 }
      },
      "SkipQuery": {
        "name": "skip",
        "in": "query",
        "required": false,
        "schema": { "type": "integer", "minimum": 0 }
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "No autenticado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "NotFound": {
        "description": "Recurso no encontrado",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
...
      "ReservationStatus": {
        "type": "string",
        "enum": ["PENDING", "CONFIRMED", "CANCELLED", "COMPLETED", "NO_SHOW"]
      },
      "ReservationSource": {
        "type": "string",
        "enum": ["WHATSAPP", "DASHBOARD", "PHONE", "WALK_IN", "API"]
      },
...
      "Customer": {
        "type": "object",
        "required": ["id", "restaurantId", "phone", "active"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "restaurantId": { "type": "string", "format": "uuid" },
          "firstName": { "type": ["string", "null"] },
          "lastName": { "type": ["string", "null"] },
          "phone": { "type": "string" },
          "email": { "type": ["string", "null"], "format": "email" },
          "preferredLanguage": { "type": "string" },
          "notes": { "type": ["string", "null"] },
          "totalReservations": { "type": "integer" },
          "lastReservationAt": { "type": ["string", "null"], "format": "date-time" },
          "active": { "type": "boolean" }
        }
      },
      "CustomerProfile": {
        "allOf": [
          { "$ref": "#/components/schemas/Customer" },
          {
            "type": "object",
            "properties": {
              "metrics": { "type": "object" },
              "reservations": { "$ref": "#/components/schemas/PaginatedReservations" }
            }
          }
        ]
      },
      "Reservation": {
        "type": "object",
        "required": ["id", "restaurantId", "customerId", "reservationStart", "reservationEnd", "guestCount", "status"],
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "restaurantId": { "type": "string", "format": "uuid" },
          "customerId": { "type": "string", "format": "uuid" },
          "reservationDate": { "type": "string", "format": "date-time" },
          "reservationStart": { "type": "string", "format": "date-time" },
          "reservationEnd": { "type": "string", "format": "date-time" },
          "guestCount": { "type": "integer", "minimum": 1 },
          "status": { "$ref": "#/components/schemas/ReservationStatus" },
          "source": { "$ref": "#/components/schemas/ReservationSource" },
          "notes": { "type": ["string", "null"] },
          "confirmationCode": { "type": ["string", "null"] },
          "customer": { "$ref": "#/components/schemas/Customer" },
          "assignedTables": { "type": "array", "items": { "type": "object" } }
        }
      },
...
      "PaginatedReservations": {
        "type": "object",
        "required": ["data", "total", "take", "skip"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Reservation" } },
          "total": { "type": "integer" },
          "take": { "type": "integer" },
          "skip": { "type": "integer" }
        }
      },
      "PaginatedCustomers": {
        "type": "object",
        "required": ["data", "total", "take", "skip"],
        "properties": {
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Customer" } },
          "total": { "type": "integer" },
          "take": { "type": "integer" },
          "skip": { "type": "integer" }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": { "type": "integer" },
          "message": {
            "oneOf": [
              { "type": "string" },
              { "type": "array", "items": { "type": "string" } }
            ]
          },
          "error": { "type": "string" }
        }
      }
    }
  }
}